fix: move withSqliteForeignKeysOff to sql utils
