+++ packages/sqlite/src/index.mjs
@@ -5,4 +5,3 @@ export const SqliteDialect = Pkg.SqliteDialect;
 export const SqliteQueryGenerator = Pkg.SqliteQueryGenerator;
 export const SqliteQueryInterface = Pkg.SqliteQueryInterface;
 export const SqliteQuery = Pkg.SqliteQuery;
-export const withSqliteForeignKeysOff = Pkg.withSqliteForeignKeysOff;
