+++ packages/sqlite/src/query-interface.internal.ts
@@ -1,10 +1,10 @@
 import type { QueryRawOptions, Sequelize, TableOrModel } from '@sequelize/core';
 import { ForeignKeyConstraintError, QueryTypes, TransactionNestMode } from '@sequelize/core';
 import { AbstractQueryInterfaceInternal } from '@sequelize/core/_non-semver-use-at-your-own-risk_/dialects/abstract/query-interface-internal.js';
+import { withSqliteForeignKeysOff } from '@sequelize/core/_non-semver-use-at-your-own-risk_/utils/sql.js';
 import type { SqliteDialect } from './dialect.js';
 import type { SqliteQueryGenerator } from './query-generator.js';
 import type { SqliteColumnsDescription } from './query-interface.types.js';
-import { withSqliteForeignKeysOff } from './sqlite-utils.js';
 
 export class SqliteQueryInterfaceInternal extends AbstractQueryInterfaceInternal {
   constructor(readonly dialect: SqliteDialect) {
