--- packages/sqlite/src/sqlite-utils.ts
@@ -1,15 +1,0 @@
-import type { QueryRawOptions, Sequelize } from '@sequelize/core';
-
-export async function withSqliteForeignKeysOff<T>(
-  sequelize: Sequelize,
-  options: QueryRawOptions | undefined,
-  cb: () => Promise<T>,
-): Promise<T> {
-  try {
-    await sequelize.queryRaw('PRAGMA foreign_keys = OFF', options);
-
-    return await cb();
-  } finally {
-    await sequelize.queryRaw('PRAGMA foreign_keys = ON', options);
-  }
-}
